% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{bootFunReg}
\alias{bootFunReg}
\title{BOOTSRTAP FOR THE FUNCTIONAL LIFECOURSE MODEL}
\usage{
bootFunReg(y, X, grid, covariates = NULL, L = 6, nBoot = 100)
}
\arguments{
\item{y}{The output vector.}

\item{X}{A n times M matrix containing the n functions evaluated at points specified in grid.}

\item{grid}{A vector of the points at which the functions are evaluated.}

\item{covariates}{A matrix of covariates.}

\item{L}{Number of B-splines basis to approximate omega.}

\item{nBoot}{Number of bootstrap samples.}
}
\value{
An list containing estimates for alpha, delta, beta, gamma, sigma2, and omega.
}
\description{
Bootsrtap of the functional regression,
}
\section{y_i = delta int X_i(t) * omega(t) dt + C_i' alpha + eps_i}{
where omega(t) is a weight function, omega(t) > 0 and in omega(t) dt = 1.
}

