% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{composite_test}
\alias{composite_test}
\title{Bayesian test for the three main hypotheses,}
\usage{
composite_test(post_w, l = 0.15, u = 0.85)
}
\arguments{
\item{post_w}{Posterior samples of the weights.}

\item{l}{lower treshold for phi  (under which we accept the accumulation model).}

\item{u}{upper treshold for phi  (above which we accept the critical model).}
}
\value{
The posterior probabilities of each three hypotheses.
}
\description{
Compute the posterior probabilities of the main different models: Accumulation, Sensitivive, Critical.
It computes the posterior distribution of phi (the range of the weights ), and compare it to a threshold.
}
\examples{
library(latFunReg)
library(dplyr)
library(GPfit)

# 1) Set-up data:
head(toy) # toy example dataset
toy2 <- toy
toy2[, 4] <- round( ( toy[, 4] - 10 ) / 30, 2 ) # i. rescale age:
y <- toy2 \%>\% group_by(id) \%>\% summarise(outcome = mean(outcome) ) \%>\% pull(outcome) # extract outcome:
t_obs <- toy2 \%>\% group_by(id) \%>\% summarise(age = list(age)) \%>\%  pull(age) # extract age
exposure <- toy2 \%>\% group_by(id) \%>\% summarise(exposure = list(exposure)) \%>\%  pull(exposure) # extract exposure

# 2) Fit Gaussian processes
gpfit <- lapply( 1:length(y), function(i) GPfit::GP_fit( t_obs[[i]], exposure[[i]] ) )

# 3) Karhunen loeve expansion:
kl <- KL_expansion( gpfit, t_obs )

# 4) Estimation (em algorithm):
est2 <- fReg_bayes(y, kl, L = 6 )
#==============
# Testing:
#==============
library(lmhyp)
library(stringi)
library(stringr)
library(purrr)
library(tidyr)
library(tibble)

# composite test: Accumulation, sensitive, critical
n_period <- 5
paritions <- cbind( 1:n_period - 1, 1:n_period ) / n_period
colnames( paritions) <- c("l", "u")
post_w <- get_post_weights(est2, paritions )
composite_test( post_w )

# Finest credible intervals:
get_FCI( post_w )
}
