% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationRLM.R
\name{fitMV_RLM}
\alias{fitMV_RLM}
\title{Multivariate RELEVANT LIFE COURSE MODEL}
\usage{
fitMV_RLM(y, x, covariates = NULL, robust = FALSE, ...)
}
\arguments{
\item{y}{The output matrix.}

\item{x}{A n times M matrix containing the n functions evaluated at points specified in grid.}

\item{robust}{If TRUE, use a t-distribution as a likelihood.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
An list containing estimates for alpha, delta, beta, gamma, sigma2, and omega.
}
\description{
Estimate the multivariate version of Madathil's regression model:
y_ij = delta_j X_it * weights_tj + C_i' alpha_j + eps_ij
}
\examples{
n <- 100
p <- 4
J <- 5
x <- matrix( rnorm(n*J), ncol = J )
delta <- c(2, 1, rep(0, p-2) )
weights <- rbind( c(0.1, 0.3, 0.6), c(0.8, 0.2, 0), matrix(0, nrow = p-2, ncol = J) )
sigma <- c(1, 0.5, rep(1, p-2))
x <- matrix( rnorm(n*J), ncol = J )
eps <- matrix( rnorm(n * p), ncol = p ) \%*\% diag(sigma)
y <- x \%*\% t(weights) \%*\% diag(delta) + eps
fitMV_RLM(y, x, warmup = 2000, iter = 4000, chains = 4)
}
