% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationRLM.R
\name{fitRLM}
\alias{fitRLM}
\title{DISCRETE RELEVANT LIFE COURSE MODEL}
\usage{
fitRLM(y, x, covariates = NULL, robust = FALSE, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{A n times M matrix containing the n functions evaluated at points specified in grid.}

\item{robust}{If TRUE, use a t-distribution as a likelihood.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
An list containing estimates for alpha, delta, beta, gamma, sigma2, and omega.
}
\description{
Estimate Madathil's regression model
y_i = delta X_ij * weights_j + C_i' alpha + eps_i
}
\examples{
n <- 100
J <- 5
x <- matrix( rnorm(n*J), ncol = J )
eps <- rnorm(n)
delta <- 2
weights <- runif(J); 
weights <- weights / sum(weights)
y <- as.numeric( delta * ( x \%*\% weights) + eps )
fitRLM(y, x, warmup = 2000, iter = 4000, chains = 4)
}
