% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRLM_additive.R
\name{fRLM}
\alias{fRLM}
\title{fRLM Function}
\usage{
fRLM(
  data,
  id,
  time,
  exposures,
  outcome,
  controls = NULL,
  L = 5,
  alpha_par = 1,
  beta_par = 1,
  ...
)
}
\arguments{
\item{data}{A tidy dataframe containing the data.}

\item{id}{A string representing the identifier variable in the dataframe.}

\item{time}{A string representing the time variable in the dataframe.}

\item{exposures}{A character vector of strings representing the exposure variables in the dataframe.}

\item{outcome}{A string representing the outcome variable in the dataframe.}

\item{controls}{A character vector of strings representing the control variables in the dataframe. Default is NULL.}

\item{L}{An integer representing the number of basis functions. Default is 5.}

\item{alpha_par}{A numeric value for the alpha hyperparameter. Default is 1.}

\item{beta_par}{A numeric value for the beta hyperparameter. Default is 1.}

\item{...}{Additional arguments passed to the \code{rstan::stan} function.}
}
\value{
A list containing the fitted model and other relevant information.
}
\description{
Wrapper function that estimates a Functional Regression Linear Model (fRLM) with a given set of predictors.
}
\details{
The function begins by scaling the time variable using a time scaler function created with \code{timeScaler_ff}.
Next, the function defines a grid of 150 points between 0 and 1 and fits a Gaussian process for each exposure
in the \code{exposures} vector using the \code{gpFit} function. The \code{gpFit} function is applied to each subject
and the predicted values are computed on the grid for each subject.
The resulting predicted values are then multiplied by a set of basis functions computed on the grid using the \code{getBasis} function,
and averaged across subjects.
The \code{data} is then aggregated by \code{id} and the means of the \code{outcome} and \code{controls} variables are computed for each subject.
The \code{data} is then passed to the \code{rstan::stan} function to fit a Bayesian regression model.
The fitted model and other relevant information is then returned as a list.
}
\examples{

# Example usage of the fRLM function
library(fRLM)
data(toy)
fRLM(data=toy,
  id = "id",
  time="age",
  exposures="exposure",
  outcome="outcome",
  warmup = 500, iter = 1000, chains = 2) # this is passed to stan

}
