% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRLM_lpfr.R
\name{fRLM_lpfr1}
\alias{fRLM_lpfr1}
\title{fRLM Function}
\usage{
fRLM_lpfr1(
  data,
  id,
  time,
  exposures,
  outcome,
  family = "gaussian",
  bounded_exposures = FALSE,
  controls = NULL,
  L = 4,
  K = 5,
  grid = seq(0, 1, l = 150),
  fit_args = list(),
  sample_args = list()
)
}
\arguments{
\item{data}{A tidy dataframe containing the data.}

\item{id}{A string representing the identifier variable in the dataframe.}

\item{time}{A string representing the time variable in the dataframe, at what time the outcome or exposures were measured.}

\item{exposures}{A character vector of strings representing the exposure variables in the dataframe. For now limited to length 1.}

\item{outcome}{A string representing the outcome variable in the dataframe.}

\item{family}{A string representing the family. Either "gaussian" or "binary".}

\item{bounded_exposures}{A boolean vector specifying whether the exposure should be bounded between 0 and 1. For now of length 1.}

\item{controls}{A character vector of strings representing the control variables in the dataframe. Default is NULL.}

\item{L}{An integer representing the number of basis functions. Default is 4.}

\item{K}{An integer representing the number of knots for each basis function. Default is K=5.}

\item{fit_args}{\code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel()}} args}

\item{sample_args}{args}
}
\description{
\code{\link[=fRLM_lpfr]{fRLM_lpfr()}} with \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}
}
