% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_bayesian.R
\name{funcReg_bayes}
\alias{funcReg_bayes}
\title{FUNCTIONAL LIFECOURSE MODEL}
\usage{
funcReg_bayes(
  y,
  X,
  L = 5,
  covariates = NULL,
  alpha_par = 1,
  beta_par = 1,
  grid = seq(0, 1, l = ncol(X)),
  ...
)
}
\arguments{
\item{y}{The output vector.}

\item{X}{A n times M matrix containing the n functions evaluated at points specified in grid.}

\item{L}{Number of B-splines basis to approximate omega.}

\item{covariates}{A matrix of covariates.}

\item{alpha_par}{Variance yperparameter of the prior of alpha.}

\item{beta_par}{Hyperparameter of the dirichlet prior of beta.}

\item{grid}{A vector of the points at which the functions are evaluated.}

\item{...}{Optional parameters to pass to the stan function.}
}
\value{
An list containing estimates for alpha, delta, beta, gamma, and sigma2.
}
\description{
Estimate a functional regression,
}
\section{y_i = delta int X_i(t) * omega(t) dt + C_i' alpha + eps_i}{
where omega(t) is a weight function, omega(t) > 0 and in omega(t) dt = 1.
}

\examples{
library(fRLM)
library(dplyr)
# parallel computing for the bayesian fit:
# library("rstan") # observe startup messages
# options(mc.cores = parallel::detectCores())
# rstan_options(auto_write = TRUE)

# 1) Setup data:
head(toy) # toy example dataset
toy2 <- toy
toy2[, 4] <- round( ( toy[, 4] - 10 ) / 30, 2 ) # i. rescale age:
y <- toy2 \%>\% group_by(id) \%>\% summarise(outcome = mean(outcome) ) \%>\% pull(outcome) # extract outcome:
t_obs <- toy2 \%>\% group_by(id) \%>\% summarise(age = list(age)) \%>\%  pull(age) # extract age
exposure <- toy2 \%>\% group_by(id) \%>\% summarise(exposure = list(exposure)) \%>\%  pull(exposure) # extract exposure

# 2) Fit gaussian processes:
gpfitList <- lapply( 1:length(y), function(i) gpFit( t_obs[[i]], exposure[[i]] ) )
grid <- seq(0,1, l = 150 )
condMu <- t( sapply( gpfitList, predict, tnew = grid ) )

# 3) Bayesian Estimation:
fitBayes <- funcReg_bayes( y, condMu, warmup = 500, iter = 1000, chains = 2 )
plot(fitBayes)
}
