% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_bayesian.R
\name{funcReg_bayes1}
\alias{funcReg_bayes1}
\title{FUNCTIONAL LIFECOURSE MODEL}
\usage{
funcReg_bayes1(
  y,
  X,
  L = 5,
  covariates = NULL,
  alpha_par = 1,
  beta_par = 1,
  grid = seq(0, 1, l = ncol(X)),
  fit_args = NULL,
  sample_args = NULL
)
}
\arguments{
\item{y}{The output vector.}

\item{X}{A n times M matrix containing the n functions evaluated at points specified in grid.}

\item{L}{Number of B-splines basis to approximate omega.}

\item{covariates}{A matrix of covariates.}

\item{alpha_par}{Variance yperparameter of the prior of alpha.}

\item{beta_par}{Hyperparameter of the dirichlet prior of beta.}

\item{grid}{A vector of the points at which the functions are evaluated.}

\item{fit_args}{args of \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}}

\item{sample_args}{args of the \code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel()}} sample method.}
}
\description{
\code{\link[=funcReg_bayes]{funcReg_bayes()}} with \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}
}
