% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcesses.R
\name{generateGaussianProcess}
\alias{generateGaussianProcess}
\title{Gaussian process simulation}
\usage{
generateGaussianProcess(
  mu = NULL,
  kernelFunc = NULL,
  sig2 = 1,
  gridLength = 100
)
}
\arguments{
\item{mu}{Mean function.}

\item{kernelFunc}{Covariance kernel function (i.e. a positive function of two arguments). Default is the exponential kernel.}

\item{sig2}{Variance of the gaussian process.}

\item{gridLength}{Number of points for simulations.}
}
\value{
A function, i.e. a realization of a Gaussian process.
}
\description{
Generate a Gaussian process on the interval \link{0,1}.
}
\examples{
  gp <- generateGaussianProcess()
  tobs <- sort(runif(5))
  Xobs <- gp(tobs)
  gpHat <- gpFit(tobs, Xobs)
  plot(gpHat)
}
