% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRLM_lpfr.R
\name{get_splines}
\alias{get_splines}
\title{Generate Spline Basis and Derivatives}
\usage{
get_splines(grid, K, range = c(0, 1), degree = 3)
}
\arguments{
\item{grid}{A numeric vector that specifies the points at which the splines and their derivatives
are to be evaluated.}

\item{K}{An integer specifying the total number of knots to be used in the spline basis.
It is to be noted that the actual number of knots used inside the function will be \code{K - 2}.}

\item{range}{A numeric vector of length 2 specifying the range over which the knots are distributed.
Defaults to c(0, 1).}

\item{degree}{An integer specifying the degree of the spline. Defaults to 3.}
}
\value{
A list containing the following elements:
- \code{psi}: The evaluated spline basis at the specified grid points.
- \code{basis}: The spline basis object as created by \code{SplineBasis}.
- \code{d_basis}: The first derivative of the spline basis.
- \code{d2psi}: The second derivative of the spline basis evaluated at the grid points.
}
\description{
This function generates a spline basis and its derivatives using a specified grid, number of knots,
and degree. It uses the \code{orthogonalsplinebasis} package for creating and manipulating spline basis.
}
\examples{
grid <- seq(0, 1, length.out = 100)
splines <- get_splines(grid, K = 5)
plot(grid, splines$psi[, 1], type = 'l')

}
