% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpfr.R
\name{lpfr_stan}
\alias{lpfr_stan}
\title{Goldsmith method}
\usage{
lpfr_stan(
  y,
  tobs,
  xobs,
  L = 4,
  K = 5,
  covariates = NULL,
  grid = seq(0, 1, l = 150),
  compiled_file = NULL,
  ...
)
}
\arguments{
\item{y:}{a vector of numeric values, the outcome, with vector elements corresponding to the units.}

\item{tobs:}{a list of time of observations, with list elements corresponding to the units}

\item{xobs:}{a list of exposures corresponding to the time observeations, with list elements corresponding to the units}

\item{covariates:}{a matrix of covariates, with rows corresponding to the units.}
}
\description{
Goldsmith method
}
