% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\docType{data}
\name{toy}
\alias{toy}
\title{toy}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
\item{Col1}{description of Col1}
\item{Col2}{description of Col2}
}
}
\usage{
toy
}
\description{
Description_of_the_Data
}
\keyword{datasets}
