% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRLM_additive.R
\name{fRLM}
\alias{fRLM}
\title{fRLM Function}
\usage{
fRLM(
  data,
  id,
  time,
  exposures,
  outcome,
  controls = NULL,
  grouping = NULL,
  L = 5,
  alpha_par = 1,
  beta_par = 1,
  ...
)
}
\arguments{
\item{data}{A tidy dataframe containing the data.}

\item{id}{A string representing the identifier variable in the dataframe.}

\item{time}{A string representing the time variable in the dataframe.}

\item{exposures}{A character vector of strings representing the exposure variables in the dataframe.}

\item{outcome}{A string representing the outcome variable in the dataframe.}

\item{controls}{A character vector of strings representing the control variables in the dataframe. Default is NULL.}

\item{grouping}{A data-frame with the same variable name for respondent identification, and as many columns as there are exposures to the random effect, in chronological order. It is understood that the time-increment is constant between exposures.}

\item{L}{An integer representing the number of basis functions. Default is 5.}

\item{alpha_par}{A numeric value for the alpha hyperparameter. Default is 1.}

\item{beta_par}{A numeric value for the beta hyperparameter. Default is 1.}

\item{...}{Additional arguments passed to the \code{rstan::stan} function.}
}
\value{
A list containing the fitted model and other relevant information.
}
\description{
Wrapper function that estimates a Functional Regression Linear Model (fRLM) with a given set of predictors.
}
\details{
The function begins by scaling the time variable using a time scaler function created with \code{timeScaler_ff}.
Next, the function defines a grid of 150 points between 0 and 1 and fits a Gaussian process for each exposure
in the \code{exposures} vector using the \code{gpFit} function. The \code{gpFit} function is applied to each subject
and the predicted values are computed on the grid for each subject.
The resulting predicted values are then multiplied by a set of basis functions computed on the grid using the \code{getBasis} function,
and averaged across subjects.
The \code{data} is then aggregated by \code{id} and the means of the \code{outcome} and \code{controls} variables are computed for each subject.
The \code{data} is then passed to the \code{rstan::stan} function to fit a Bayesian regression model.
The fitted model and other relevant information is then returned as a list.
}
\examples{

# Minimal example usage of the fRLM function
library(fRLM)
data(toy)
output <- fRLM(data=toy,
  id = "id",
  time="age",
  exposures="exposure",
  outcome="outcome",
  warmup = 500, iter = 1000, chains = 2) # this is passed to stan

# Example with more exposures and random effects


# Example usage of the fRLM function

library(fRLM)
set.seed(1234)
data(toy)

toy2 <- toy
# Add an unrelated exposure
toy2$exposure2 <- round(rnorm(nrow(toy)),2)
# Add a grouping
grouping <- unique(toy2$id) \%>\% as_tibble() \%>\% rename(id=value)
grouping <- grouping \%>\% mutate(group = 1 + (runif(n()) <0.5)*1)

# Modify the outcome to add random effects
add_to_outcome <- grouping \%>\% mutate(to_add=ifelse(group==1, -0.5, +0.5)) \%>\% dplyr::select(id, to_add)
toy2 <- toy2 \%>\% left_join(add_to_outcome) \%>\% mutate(outcome = outcome + to_add) \%>\% select(-to_add)


output <- fRLM(data=toy2,
               id = "id",
               time="age",
               exposures=c("exposure", "exposure2"),
               grouping = grouping,
               outcome="outcome",
               warmup = 1000, iter = 2000, chains = 2) # this is passed to stan


samples <- rstan::extract(output$fit, permuted = TRUE)
# Create confidence intervals for delta (effect size of exposures)
apply(output$delta, 2, function(x) quantile(x, c(0.025, 0.975)))

plot(output)
}
